requirements.txt
# Core Flask
flask==3.0.0
gunicorn==21.2.0

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Google Sheets/Drive (lightweight)
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# Data processing (for automation components)
pandas>=2.0.0
numpy>=1.24.0
